requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.19.0
transformers>=4.30.0
pillow>=9.5.0
numpy>=1.24.0
rembg>=2.0.0
gradio>=3.35.0
accelerate>=0.20.0
